old/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.7
redis==5.0.4
SQLAlchemy==2.0.41
asyncpg==0.30.0
alembic==1.16.2
passlib==1.7.4
bcrypt==4.3.0
email_validator==2.2.0
python-jose[cryptography]==3.5.0
pytest==8.4.1
pytest-asyncio==1.0.0
ruff==0.12.0
black==25.1.0
fastapi-cli==0.0.7
prometheus-fastapi-instrumentator==7.1.0
sentry-sdk[fastapi]==2.35.0